EndOfSection18

+++ client/src/app/messages/messages.component.ts
@@ -1,6 +1,8 @@
+import { ChangeDetectorRef } from '@angular/core';
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -15,8 +17,8 @@ export class MessagesComponent implements OnInit {
   pageNumber = 1;
   pageSize = 5;
   loading = false;
-
-  constructor(private messageService: MessageService) { }
+  
+  constructor(private messageService: MessageService, private cdr: ChangeDetectorRef, private confirmService: ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -32,13 +34,13 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage(id: number) {
-    // this.confirmService.confirm('Confirm delete message', 'This cannot be undone').subscribe(result => {
-    //   if (result) {
+    this.confirmService.confirm('Confirm delete message', 'This cannot be undone').subscribe(result => {
+      if (result) {
         this.messageService.deleteMessage(id).subscribe(() => {
           this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
         })
-    //   }
-    // })
+      }
+    })
 
   }
